requirements.txt
fastapi
sqlmodel
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic-settings
httpx
pytest
python-jose[cryptography]
google-generativeai
groq
clerk-backend-api
